Update client.go
Client.go is modified with the new specific. The old lines of code where commented in case someone needs them. I've added a go routine to check if the switch is disconnected. In the end i've changed the signature of the Run function with the addition of a Context and a *grpc.ClientConn